requirements.txt
Django==6.0
django-filter==25.1
djangorestframework==3.16.1
whitenoise==6.11.0
gunicorn==23.0.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
dj-database-url==3.0.1
django-cors-headers==4.9.0
python-dotenv==1.2.1
django-countries==7.6.1
celery==5.5.3
ipython==9.3.0
drf-yasg==1.21.10
django-extensions==4.1
django-redis==6.0.0
django-ipware==7.0.1
channels==4.3.2
requests==2.32.5
geoip2==5.2.0